put RecyclerViews as views in BottomNavFrag.kt remove unused fragments
