+++ .idea/misc.xml
@@ -7,7 +7,7 @@
         <entry key="app/src/main/res/drawable/ic_baseline_arrow_back_24.xml" value="0.212" />
         <entry key="app/src/main/res/drawable/ic_baseline_close_24.xml" value="0.19947916666666668" />
         <entry key="app/src/main/res/drawable/ic_baseline_play_circle_filled_24.xml" value="0.2755" />
-        <entry key="app/src/main/res/layout-land/fragment_bottom_nav.xml" value="0.21458333333333332" />
+        <entry key="app/src/main/res/layout-land/fragment_bottom_nav.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_image_viewer.xml" value="0.34479166666666666" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.16354166666666667" />
         <entry key="app/src/main/res/layout/activity_video_player.xml" value="0.17395833333333333" />
@@ -15,6 +15,8 @@
         <entry key="app/src/main/res/layout/fragment_album_detail.xml" value="0.2046875" />
         <entry key="app/src/main/res/layout/fragment_bin.xml" value="0.34479166666666666" />
         <entry key="app/src/main/res/layout/fragment_bottom_nav.xml" value="0.2777777777777778" />
+        <entry key="app/src/main/res/layout/fragment_grid_album.xml" value="0.1828125" />
+        <entry key="app/src/main/res/layout/fragment_grid_item.xml" value="0.1828125" />
         <entry key="app/src/main/res/layout/fragment_video_player.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_view_pager.xml" value="0.25" />
         <entry key="app/src/main/res/layout/list_grid_media_item_holder.xml" value="0.22870370370370371" />
