+++ app/src/main/res/layout-land/fragment_bottom_nav.xml
@@ -22,13 +22,21 @@
             app:titleCentered="true" />
     </com.google.android.material.appbar.AppBarLayout>
 
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/fcvBottomNav"
-        android:name="com.example.gallery.ui.GridItemFrag"
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rvItems"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        tools:context=".ui.GridItemFrag"
+        app:layout_behavior="@string/appbar_scrolling_view_behavior" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/rvAlbums"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:transitionGroup="true"
+        android:visibility="gone"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:layout="@layout/fragment_grid_item" />
+        tools:context=".ui.GridAlbumFrag" />
 
     <com.google.android.material.navigationrail.NavigationRailView
         android:id="@+id/bnvMain"
