--- app/src/main/res/layout/fragment_grid_album.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.recyclerview.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/rvAlbum"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.GridAlbumFrag"
-    app:layoutManager="GridLayoutManager"
-    android:transitionGroup="true"/>
